chore: MainHeader.js 반응형 업데이트
